# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Glitch API",
        "description": "The API for connecting to Glitch",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/competitions/{uuid}/users": {
            "get": {
                "tags": [
                    "Competitions Route"
                ],
                "summary": "List all the users associated with a competition.",
                "description": "List all the users associated with a competition.",
                "operationId": "competitionUserList",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the competition",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CompetitionUser"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The competition does not exist.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "The competition does not exist."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Competitions Route"
                ],
                "summary": "Associate a new users with the competition.",
                "description": "Associate a new users with the competition.",
                "operationId": "createCompetitionUser",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the competition",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompetitionUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompetitionUser"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The competition does not exist.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "The competition does not exist."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied to competition.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "Access denied to competition."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errors when creating user"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/competitions/{uuid}/users/{user_id}": {
            "get": {
                "tags": [
                    "Competitions Route"
                ],
                "summary": "Show a single user by its ID.",
                "description": "Show a single user by its ID.",
                "operationId": "showCompetitionUser",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the competition",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "UUID of the user.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompetitionUser"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Competitions Route"
                ],
                "summary": "Update the user associated with competition.",
                "description": "Update the user associated with competition.",
                "operationId": "updateCompetitionUser",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the competition.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "UUID of the user.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompetitionUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompetitionUser"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The competition does not exist.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "Access denined to venue."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied to competition.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "Access denined to user."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Competitions Route"
                ],
                "summary": "Remove the associated user from the competition.",
                "description": "Remove the associated user from the competition.",
                "operationId": "removeCompetitionUser",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the competition.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "UUID of the user.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "The competition does not exist.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "Access denined to competition."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied to team.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "Access denined to competition."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CompetitionUser": {
                "title": "Competition User Model",
                "description": "Manage a user relationship with a competition.",
                "required": [
                    "competition_id",
                    "user_id"
                ],
                "properties": {
                    "competition_id": {
                        "description": "The id of the competition.",
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "user_id": {
                        "description": "The id of the user associated with the competition.",
                        "type": "string",
                        "format": "uuid"
                    },
                    "checked_in": {
                        "description": "If the user has checked-in.",
                        "type": "boolean"
                    },
                    "waiver_signed": {
                        "description": "If the waiver has been signed for the user.",
                        "type": "boolean"
                    },
                    "entry_fee_paid": {
                        "description": "If the waiver has been signed for the user.",
                        "type": "boolean"
                    },
                    "status": {
                        "description": "The current status of the user in the competition.",
                        "type": "integer",
                        "enum": [
                            "1",
                            "2",
                            "3",
                            "4",
                            "5",
                            "6",
                            "7",
                            "8",
                            "9"
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    }
}